Move file s3 operations to CRUD module
